+++ src/cjs/entry.js
@@ -1,8 +1,6 @@
 "use strict";
 /* Imports for webpack */
 
-// import AlloyFinger, { TouchMoveEvent, TouchPinchEvent } from "alloyfinger-typescript/src/js/index";
-
 globalThis.UIDGenerator = require("./UIDGenerator.js").UIDGenerator;
 globalThis.VertexAttr = require("./VertexAttr.js").VertexAttr;
 globalThis.VertexListeners = require("./VertexListeners.js").VertexListeners;
@@ -23,8 +21,7 @@ globalThis.PBImage = require("./PBImage.js").PBImage;
 globalThis.MouseHandler = require("./MouseHandler.js").MouseHandler;
 globalThis.KeyHandler = require("./KeyHandler.js").KeyHandler;
 globalThis.drawutils = require("./draw.js").drawutils;
-// globalThis.drawutilsgl = require("./drawgl.js").drawutilsgl;
-// globalThis.drawutilsgl = {};
-globalThis.drawutilssvg = require("./utils/helpers/drawutilssvg.js").drawutilssvg;
+// globalThis.drawutilsgl = require("./drawgl.js").drawutilsgl;  // Unfinished
+globalThis.drawutilssvg = require("./drawutilssvg.js").drawutilssvg;
 globalThis.geomutils = require("./geomutils.js").geomutils;
 globalThis.PlotBoilerplate = require("./PlotBoilerplate.js").PlotBoilerplate;