+++ src/esm/draw.js
@@ -39,7 +39,7 @@
  **/
 import { CubicBezierCurve } from "./CubicBezierCurve";
 import { Vertex } from "./Vertex";
-import { drawutilssvg } from "./utils/helpers/drawutilssvg";
+import { drawutilssvg } from "./drawutilssvg";
 // Todo: rename this class to Drawutils?
 /**
  * @classdesc A wrapper class for basic drawing operations.
@@ -64,7 +64,6 @@ export class drawutils {
         this.scale = new Vertex(1, 1);
         this.fillShapes = fillShapes;
     }
-    ;
     /**
      * Called before each draw cycle.
      * @param {UID=} uid - (optional) A UID identifying the currently drawn element(s).
@@ -72,7 +71,6 @@ export class drawutils {
     beginDrawCycle(renderTime) {
         // NOOP
     }
-    ;
     /**
      * This method shouled be called each time the currently drawn `Drawable` changes.
      * It is used by some libraries for identifying elemente on re-renders.
@@ -84,7 +82,6 @@ export class drawutils {
     setCurrentId(uid) {
         // NOOP
     }
-    ;
     /**
      * This method shouled be called each time the currently drawn `Drawable` changes.
      * Determine the class name for further usage here.
@@ -96,7 +93,6 @@ export class drawutils {
     setCurrentClassName(className) {
         // NOOP
     }
-    ;
     /**
      * Draw the line between the given two points with the specified (CSS-) color.
      *
@@ -119,7 +115,6 @@ export class drawutils {
         this.ctx.stroke();
         this.ctx.restore();
     }
-    ;
     /**
      * Draw a line and an arrow at the end (zB) of the given line with the specified (CSS-) color.
      *
@@ -148,7 +143,6 @@ export class drawutils {
         this._fillOrDraw(color);
         this.ctx.restore();
     }
-    ;
     /**
      * Draw an image at the given position with the given size.<br>
      * <br>
@@ -175,7 +169,6 @@ export class drawutils {
         this.offset.x + position.x * this.scale.x, this.offset.y + position.y * this.scale.y, size.x * this.scale.x, size.y * this.scale.y);
         this.ctx.restore();
     }
-    ;
     /**
      * Draw a rectangle.
      *
@@ -198,7 +191,6 @@ export class drawutils {
         this._fillOrDraw(color);
         this.ctx.restore();
     }
-    ;
     // +---------------------------------------------------------------------------------
     // | This is the final helper function for drawing and filling stuff. It is not
     // | intended to be used from the outside.
@@ -221,7 +213,6 @@ export class drawutils {
             this.ctx.stroke();
         }
     }
-    ;
     /**
      * Draw the given (cubic) bézier curve.
      *
@@ -251,7 +242,6 @@ export class drawutils {
         this._fillOrDraw(color);
         this.ctx.restore();
     }
-    ;
     /**
      * Draw the given (quadratic) bézier curve.
      *
@@ -275,7 +265,6 @@ export class drawutils {
         this._fillOrDraw(color);
         this.ctx.restore();
     }
-    ;
     /**
      * Draw the given (cubic) Bézier path.
      *
@@ -312,7 +301,6 @@ export class drawutils {
         this._fillOrDraw(color);
         this.ctx.restore();
     }
-    ;
     /**
      * Draw the given handle and handle point (used to draw interactive Bézier curves).
      *
@@ -328,10 +316,9 @@ export class drawutils {
     handle(startPoint, endPoint) {
         // Draw handles
         // (No need to save and restore here)
-        this.point(startPoint, 'rgb(0,32,192)');
-        this.square(endPoint, 5, 'rgba(0,128,192,0.5)');
+        this.point(startPoint, "rgb(0,32,192)");
+        this.square(endPoint, 5, "rgba(0,128,192,0.5)");
     }
-    ;
     /**
      * Draw a handle line (with a light grey).
      *
@@ -344,9 +331,8 @@ export class drawutils {
      */
     handleLine(startPoint, endPoint) {
         // Draw handle lines
-        this.line(startPoint, endPoint, 'rgb(192,192,192)');
+        this.line(startPoint, endPoint, "rgb(192,192,192)");
     }
-    ;
     /**
      * Draw a 1x1 dot with the specified (CSS-) color.
      *
@@ -367,7 +353,6 @@ export class drawutils {
         this._fillOrDraw(color);
         this.ctx.restore();
     }
-    ;
     /**
      * Draw the given point with the specified (CSS-) color and radius 3.
      *
@@ -386,7 +371,6 @@ export class drawutils {
         this.ctx.lineWidth = 1;
         this._fillOrDraw(color);
     }
-    ;
     /**
      * Draw a circle with the specified (CSS-) color and radius.<br>
      * <br>
@@ -408,22 +392,21 @@ export class drawutils {
         this.ctx.lineWidth = lineWidth || 1;
         this._fillOrDraw(color);
     }
-    ;
     /**
-     * Draw a circular arc (section of a circle) with the given CSS color.
-     *
-     * @method circleArc
-     * @param {Vertex} center - The center of the circle.
-     * @param {number} radius - The radius of the circle.
-     * @param {number} startAngle - The angle to start at.
-     * @param {number} endAngle - The angle to end at.
-     * @param {string=#000000} color - The CSS color to draw the circle with.
-     * @param {number=1} lineWidth - The line width to use
-     // * @param {boolean=false} options.asSegment - If `true` then no beginPath and no draw will be applied (as part of larger path).
-     * @return {void}
-     * @instance
-     * @memberof drawutils
-     */
+       * Draw a circular arc (section of a circle) with the given CSS color.
+       *
+       * @method circleArc
+       * @param {Vertex} center - The center of the circle.
+       * @param {number} radius - The radius of the circle.
+       * @param {number} startAngle - The angle to start at.
+       * @param {number} endAngle - The angle to end at.
+       * @param {string=#000000} color - The CSS color to draw the circle with.
+       * @param {number=1} lineWidth - The line width to use
+       // * @param {boolean=false} options.asSegment - If `true` then no beginPath and no draw will be applied (as part of larger path).
+       * @return {void}
+       * @instance
+       * @memberof drawutils
+       */
     circleArc(center, radius, startAngle, endAngle, color, lineWidth, options) {
         if (!options || !options.asSegment) {
             this.ctx.beginPath();
@@ -432,10 +415,9 @@ export class drawutils {
         if (!options || !options.asSegment) {
             // this.ctx.closePath();
             this.ctx.lineWidth = lineWidth || 1;
-            this._fillOrDraw(color || '#000000');
+            this._fillOrDraw(color || "#000000");
         }
     }
-    ;
     /**
      * Draw an ellipse with the specified (CSS-) color and thw two radii.
      *
@@ -451,7 +433,7 @@ export class drawutils {
      * @memberof drawutils
      */
     ellipse(center, radiusX, radiusY, color, lineWidth, rotation) {
-        if (typeof rotation === 'undefined') {
+        if (typeof rotation === "undefined") {
             rotation = 0.0;
         }
         this.ctx.beginPath();
@@ -460,7 +442,6 @@ export class drawutils {
         this.ctx.lineWidth = lineWidth || 1;
         this._fillOrDraw(color);
     }
-    ;
     /**
      * Draw square at the given center, size and with the specified (CSS-) color.<br>
      * <br>
@@ -482,7 +463,6 @@ export class drawutils {
         this.ctx.lineWidth = lineWidth || 1;
         this._fillOrDraw(color);
     }
-    ;
     /**
      * Draw a grid of horizontal and vertical lines with the given (CSS-) color.
      *
@@ -516,7 +496,6 @@ export class drawutils {
         this.ctx.stroke();
         this.ctx.closePath();
     }
-    ;
     /**
      * Draw a raster of crosshairs in the given grid.<br>
      *
@@ -551,7 +530,6 @@ export class drawutils {
         this.ctx.closePath();
         this.ctx.restore();
     }
-    ;
     /**
      * Draw a diamond handle (square rotated by 45°) with the given CSS color.
      *
@@ -577,7 +555,6 @@ export class drawutils {
         this.ctx.lineWidth = 1;
         this._fillOrDraw(color);
     }
-    ;
     /**
      * Draw a square handle with the given CSS color.<br>
      * <br>
@@ -600,7 +577,6 @@ export class drawutils {
         this.ctx.lineWidth = 1;
         this._fillOrDraw(color);
     }
-    ;
     /**
      * Draw a circle handle with the given CSS color.<br>
      * <br>
@@ -624,7 +600,6 @@ export class drawutils {
         this.ctx.lineWidth = 1;
         this._fillOrDraw(color);
     }
-    ;
     /**
      * Draw a crosshair with given radius and color at the given position.<br>
      * <br>
@@ -651,7 +626,6 @@ export class drawutils {
         this.ctx.closePath();
         this.ctx.restore();
     }
-    ;
     /**
      * Draw a polygon.
      *
@@ -666,7 +640,6 @@ export class drawutils {
     polygon(polygon, color, lineWidth) {
         this.polyline(polygon.vertices, polygon.isOpen, color, lineWidth);
     }
-    ;
     /**
      * Draw a polygon line (alternative function to the polygon).
      *
@@ -689,20 +662,20 @@ export class drawutils {
         for (var i = 0; i < vertices.length; i++) {
             this.ctx.lineTo(this.offset.x + vertices[i].x * this.scale.x, this.offset.y + vertices[i].y * this.scale.y);
         }
-        if (!isOpen) // && vertices.length > 2 )
+        if (!isOpen)
+            // && vertices.length > 2 )
             this.ctx.closePath();
         this._fillOrDraw(color);
         this.ctx.closePath();
         this.ctx.setLineDash([]);
         this.ctx.restore();
     }
-    ;
     text(text, x, y, options) {
         options = options || {};
         this.ctx.save();
         x = this.offset.x + x * this.scale.x;
         y = this.offset.y + y * this.scale.y;
-        const color = options.color || 'black';
+        const color = options.color || "black";
         if (this.fillShapes) {
             this.ctx.fillStyle = color;
             this.ctx.fillText(text, x, y);
@@ -713,7 +686,6 @@ export class drawutils {
         }
         this.ctx.restore();
     }
-    ;
     /**
      * Draw a non-scaling text label at the given position.
      *
@@ -732,9 +704,9 @@ export class drawutils {
     label(text, x, y, rotation, color) {
         this.ctx.save();
         this.ctx.translate(x, y);
-        if (typeof rotation != 'undefined')
+        if (typeof rotation != "undefined")
             this.ctx.rotate(rotation);
-        this.ctx.fillStyle = color || 'black';
+        this.ctx.fillStyle = color || "black";
         if (this.fillShapes) {
             this.ctx.fillText(text, 0, 0);
         }
@@ -743,7 +715,6 @@ export class drawutils {
         }
         this.ctx.restore();
     }
-    ;
     /**
      * Draw an SVG-like path given by the specified path data.
      *
@@ -771,7 +742,6 @@ export class drawutils {
             this.ctx.stroke(new Path2D(d.join(" ")));
         }
     }
-    ;
     /**
      * Due to gl compatibility there is a generic 'clear' function required
      * to avoid accessing the context object itself directly.
@@ -785,6 +755,5 @@ export class drawutils {
         this.ctx.fillStyle = color;
         this.ctx.fillRect(0, 0, this.ctx.canvas.width, this.ctx.canvas.height);
     }
-    ;
 }
 //# sourceMappingURL=draw.js.map