+++ src/esm/entry.js
@@ -1,9 +1,6 @@
 "use strict";
 /* Imports for webpack */
 
-// import path from "path";
-// import AlloyFinger, { TouchMoveEvent, TouchPinchEvent } from "alloyfinger-typescript/src/js/index";
-
 globalThis.UIDGenerator = require("./UIDGenerator.js").geomutils;
 globalThis.VertexAttr = require("./VertexAttr.js").VertexAttr;
 globalThis.VertexListeners = require("./VertexListeners.js").VertexListeners;
@@ -25,8 +22,6 @@ globalThis.MouseHandler = require("./MouseHandler.js").MouseHandler;
 globalThis.KeyHandler = require("./KeyHandler.js").KeyHandler;
 globalThis.drawutils = require("./draw.js").drawutils;
 // globalThis.drawutilsgl = require("./drawgl.js").drawutilsgl;
-// globalThis.drawutilsgl = {};
-globalThis.drawutilssvg = require("./utils/helpers/drawutilssvg.js").drawutilssvg;
+globalThis.drawutilssvg = require("./drawutilssvg.js").drawutilssvg;
 globalThis.geomutils = require("./geomutils.js").geomutils;
 globalThis.PlotBoilerplate = require("./PlotBoilerplate.js").PlotBoilerplate;
-