+++ src/ts/PlotBoilerplate.ts
@@ -71,7 +71,8 @@
  * @modified 2021-02-18 Adding `adjustOffset(boolean)` function.
  * @modified 2021-03-01 Updated the `PlotBoilerplate.draw(...)` function: ellipses are now rotate-able.
  * @modified 2021-03-03 Added the `VEllipseSector` drawable.
- * @version  1.13.2
+ * @modified 2021-03-29 Clearing `currentClassName` and `currentId` after drawing each drawable.
+ * @version  1.13.3
  *
  * @file PlotBoilerplate
  * @fileoverview The main class.
@@ -84,7 +85,7 @@ import { GUI } from "dat.gui";
 
 import { drawutils } from "./draw";
 import { drawutilsgl } from "./drawgl";
-import { drawutilssvg } from "./utils/helpers/drawutilssvg";
+import { drawutilssvg } from "./drawutilssvg";
 import { BezierPath } from "./BezierPath";
 import { Bounds } from "./Bounds";
 import { Circle } from "./Circle";
@@ -1263,6 +1264,10 @@ export class PlotBoilerplate {
     } else {
       console.error("Cannot draw object. Unknown class.");
     }
+    draw.setCurrentClassName(null);
+    draw.setCurrentId(null);
+    fill.setCurrentClassName(null);
+    fill.setCurrentId(null);
   }
 
   /**